# cql3: result_set, selector: change value type to managed_bytes_opt