--- a/cql3/statements/raw/select_statement.hh
+++ b/cql3/statements/raw/select_statement.hh
@@ -75,7 +75,7 @@ public:
     template<typename T>
     using compare_fn = std::function<bool(const T&, const T&)>;
 
-    using result_row_type = std::vector<bytes_opt>;
+    using result_row_type = std::vector<managed_bytes_opt>;
     using ordering_comparator_type = compare_fn<result_row_type>;
 private:
     using prepared_orderings_type = std::vector<std::pair<const column_definition*, ordering>>;