--- a/cql3/untyped_result_set.hh
+++ b/cql3/untyped_result_set.hh
@@ -41,13 +41,12 @@ class metadata;
 
 class untyped_result_set_row {
 public:
-    using view_type = query::result_bytes_view;
+    using view_type = bytes_view;
     using opt_view_type = std::optional<view_type>;
-    using view_holder = std::variant<std::monostate, view_type, bytes>;
 private:
     friend class untyped_result_set;
     using index_map = std::unordered_map<std::string_view, size_t>;
-    using data_views = std::vector<view_holder>;
+    using data_views = std::vector<bytes_opt>;
 
     const index_map& _name_to_index;
     const cql3::metadata& _metadata;
@@ -62,7 +61,7 @@ public:
     bool has(std::string_view) const;
     view_type get_view(std::string_view name) const;
     bytes get_blob(std::string_view name) const {
-        return get_view(name).linearize();
+        return to_bytes(get_view(name));
     }
     managed_bytes get_blob_fragmented(std::string_view name) const {
         return managed_bytes(get_view(name));