--- a/serializer.cc
+++ b/serializer.cc
@@ -13,3 +13,26 @@ namespace ser {
 logging::logger serlog("serializer");
 
 } // namespace ser
+
+namespace utils {
+
+managed_bytes_view
+buffer_view_to_managed_bytes_view(ser::buffer_view<bytes_ostream::fragment_iterator> bv) {
+    auto impl = bv.extract_implementation();
+    return build_managed_bytes_view_from_internals(
+            impl.current,
+            impl.next.extract_implementation().current_chunk,
+            impl.size
+    );
+}
+
+managed_bytes_view_opt
+buffer_view_to_managed_bytes_view(std::optional<ser::buffer_view<bytes_ostream::fragment_iterator>> bvo) {
+    if (!bvo) {
+        return std::nullopt;
+    }
+    return buffer_view_to_managed_bytes_view(*bvo);
+}
+
+
+} // namespace utils