@nyh 2023-04-24
> To solve the problem, we harmonize the data types to managed_bytes_opt (#13216 notwithstanding). 

I didn't understand this comment. #13216 recommends that "we should forbid the use of managed types outside of the cache and memtable data trees.". Doesn't this change break this recommendation? Why?

Perhaps we need a new type for non-contiguous strings of bytes which are nevertheless allocated using the normal allocator? We do have, by the way, several such types already, maybe it's time to unify some of them.... We have rjson::chunked_content (which is an alias for std::vector<temporary_buffer<char>>), we have std::deque, utils::chunked_vector, net::packet, and I wouldn't be surprised if we have more :-(