@michoecho 2023-04-24
> and I wouldn't be surprised if we have more

Also `bytes_ostream` and `ser::buffer_view`.