@avikivity 2023-04-24
> > To solve the problem, we harmonize the data types to managed_bytes_opt (#13216 notwithstanding).
> 
> I didn't understand this comment. #13216 recommends that "we should forbid the use of managed types outside of the cache and memtable data trees.". Doesn't this change break this recommendation? Why?

The recommendation is already broken, we use managed_bytes_opt extensively. This series moves in the wrong direction in that it adds yet more users, but also in the right direction: it reduces uses of things that aren't managed_bytes_opt, and when we fully standardize on managed_bytes_opt, it will be easier to replace it.

Fixing #13216 is a large undertaking. I don't think we can stop progress until it's fixed.


> 
> Perhaps we need a new type for non-contiguous strings of bytes which are nevertheless allocated using the normal allocator? We do have, by the way, several such types already, maybe it's time to unify some of them.... We have rjson::chunked_content (which is an alias for std::vector<temporary_buffer>), we have std::deque, utils::chunked_vector, net::packet, and I wouldn't be surprised if we have more :-(

The obvious candidate is bytes_ostream. But this is difficult, because things like atomic_cell use managed_bytes_opt and they exist both in memtable/cache and outside. So we'll need to be very careful here.