@nyh 2023-04-24
> The recommendation is already broken, we use managed_bytes_opt extensively. This series moves in the wrong direction in that it adds yet more users, but also in the right direction: it reduces uses of things that aren't managed_bytes_opt, and when we fully standardize on managed_bytes_opt, it will be easier to replace it.
> 
> Fixing #13216 is a large undertaking. I don't think we can stop progress until it's fixed.

In one of the comments you explained that bytes_ostream uses the same underlying data type as managed bytes, so I wonder if we can't:

1. Create a class "fragmented_bytes_view" (or similar name) that gives you a view into either bytes_ostream or managed_bytes - which internally have the same data structure (supposedly).
2. In the code that doesn't use LSA, use bytes_ostream instead of managed_bytes, and the existing code will use managed_bytes - but both will be usable as a fragmented_bytes_view in the new function.