@tgrabiec 2023-04-24
> Fixing #13216 is a large undertaking. I don't think we can stop progress until it's fixed.

It can be easily fixed by having a wrapper which destroys the underlying managed_bytes under `with_allocator(standard_allocator(), ...)`. Views can stay the same.