@avikivity 2023-04-24
> > Fixing #13216 is a large undertaking. I don't think we can stop progress until it's fixed.
> 
> It can be easily fixed by having a wrapper which destroys the underlying managed_bytes under `with_allocator(standard_allocator(), ...)`. Views can stay the same.


My plan was to have a base type with a private destructor, then managed_bytes (with allocator-aware destructor) and bytes_ostream, or maybe unmanaged_bytes, or whatever, with standard destructor.