@avikivity 2023-04-24
v2:

- explain immediately-invoked lambda trick in
  managed_bytes_view::with_linearized() changelog

- documented copies in to_bytes_opt(), to_managed_bytes_opt()

- unregressed alternator::executor::describe_item() for cells
  smaller than 128kB.