@avikivity 2023-05-07
v3:
 - rebased and retested